Suggested patch for #502 [#502 state:patch responsible:paulcc]
Paperclip catches its exceptions and converts them into an errors
list. This patch checks the attachment error list as a validation
step, and raises an exception if a problem is found.

However: the paperclip errors aren't very informative, so the exception
message suggests checking the imagemagick installation and the original
image file.